Add files via upload
Jinyhup, I modified your code some. I am trying to add a feature that will make a csv file from the questions and answers that a user inputs. I am also trying to make it so that a user can upload a csv file into the program form which it can pull questions and answers from the file. I'm having trouble getting it all to work together.